Durama/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
Authlib==1.6.3
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==45.0.7
cycler==0.12.1
daphne==4.2.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.6
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-redis==6.0.0
django-silk==5.4.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dparse==0.6.4
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
fqdn==1.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-vision==3.10.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gprof2dot==2025.4.14
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
msgpack==1.1.1
mssql-django==1.5
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook_shim==0.2.4
numpy==2.2.3
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydotplus==2.0.2
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==25.1.0
pyparsing==3.2.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==3.3.0
python3-openid==3.2.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
redis==6.4.0
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safety==3.6.1
safety-schemas==0.0.14
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.6.1
typer==0.17.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
Unidecode==1.4.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.29.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
zope.interface==7.2
